+++ src/main/java/es/in2/issuer/backend/shared/domain/service/impl/VerifierServiceImpl.java
@@ -8,6 +8,7 @@ import com.nimbusds.jose.crypto.RSASSAVerifier;
 import com.nimbusds.jose.jwk.*;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
+import es.in2.issuer.backend.shared.domain.exception.JWTParsingException;
 import es.in2.issuer.backend.shared.domain.exception.JWTVerificationException;
 import es.in2.issuer.backend.shared.domain.exception.TokenFetchException;
 import es.in2.issuer.backend.shared.domain.exception.WellKnownInfoFetchException;
@@ -66,6 +67,7 @@ public class VerifierServiceImpl implements VerifierService {
                 .flatMap(metadata -> fetchJWKSet(metadata.jwksUri()))
                 .flatMap(jwkSet -> {
                     try {
+                        //todo usar jwtservice.parseJWT?
                         SignedJWT signedJWT = SignedJWT.parse(accessToken);
                         JWTClaimsSet claims = signedJWT.getJWTClaimsSet();
 
@@ -88,7 +90,7 @@ public class VerifierServiceImpl implements VerifierService {
                         return Mono.empty(); // Valid token
                     } catch (ParseException | JOSEException e) {
                         log.error("Error parsing or verifying JWT", e);
-                        return Mono.error(new JWTVerificationException("Error parsing or verifying JWT"));
+                        return Mono.error(new JWTParsingException("Error parsing or verifying JWT"));
                     }
                 });
     }